Refactor BorsService to make it easier to test
