+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod command;
 pub mod github;
 pub mod permissions;
+pub mod service;
 
 mod config;
 pub(crate) mod tests;
